backend/requirements.txt
# AI Data Analysis Agent - Backend dependencies
# Python 3.10+

# Web
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Data
pandas>=2.1.0
numpy>=1.26.0
openpyxl>=3.1.0

# Analysis & ML
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0

# LLM (optional: install only what you use)
openai>=1.12.0
google-generativeai>=0.5.0

# Config & validation
pydantic>=2.5.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
